attempting to exclude entire file
