+++ kotlin-spark-api/3.2/src/test/kotlin/org/jetbrains/kotlinx/spark/api/StreamingTest.kt
@@ -64,111 +64,6 @@ class StreamingTest : ShouldSpec({
 
     context("streaming") {
 
-        context("kafka") {
-
-//            val kafka = install(
-//                TestContainerExtension(KafkaContainer(DockerImageName.parse("confluentinc/cp-kafka:6.2.1")))
-//            ) {
-//                withEmbeddedZookeeper()
-//            }
-//
-//            should("support kafka streams") {
-//                val topic1 = "test1"
-//                val topic2 = "test2"
-//
-//                val producer = kafka.createStringStringProducer()
-//                producer.send(ProducerRecord(topic1, "Hello this is a test test test"))
-//                producer.send(ProducerRecord(topic1, "Hello this is a test test test"))
-//                producer.send(ProducerRecord(topic1, "Hello this is a test test test"))
-//                producer.send(ProducerRecord(topic1, "Hello this is a test test test"))
-//                producer.send(ProducerRecord(topic2, "This is also also a test test something"))
-//                producer.send(ProducerRecord(topic2, "This is also also a test test something"))
-//                producer.send(ProducerRecord(topic2, "This is also also a test test something"))
-//                producer.send(ProducerRecord(topic2, "This is also also a test test something"))
-//                producer.close()
-//
-//                val consumer = kafka.createStringStringConsumer {
-//                    this[ConsumerConfig.MAX_POLL_RECORDS_CONFIG] = 1
-//                }
-//
-//                consumer.subscribe(listOf(topic1))
-//                val records = consumer.poll(Duration.ofSeconds(100))
-//                records.shouldHaveSize(4)
-//                records.print()
-
-            val port = 9092
-            val broker = "localhost:$port"
-            val topic1 = "test1"
-            val topic2 = "test2"
-            val kafkaListener = EmbeddedKafkaListener(port)
-            listener(kafkaListener)
-
-            should("support kafka streams") {
-                val producer = kafkaListener.stringStringProducer()
-                producer.send(ProducerRecord(topic1, "Hello this is a test test test"))
-                producer.send(ProducerRecord(topic2, "This is also also a test test something"))
-                producer.close()
-
-                withSparkStreaming(
-                    batchDuration = Durations.seconds(2),
-                    appName = "KotlinDirectKafkaWordCount",
-                    timeout = 1000L,
-                ) {
-
-                    val kafkaParams: Map<String, Serializable> = mapOf(
-                        ConsumerConfig.BOOTSTRAP_SERVERS_CONFIG to broker,
-                        ConsumerConfig.GROUP_ID_CONFIG to "consumer-group",
-                        ConsumerConfig.KEY_DESERIALIZER_CLASS_CONFIG to StringDeserializer::class.java,
-                        ConsumerConfig.VALUE_DESERIALIZER_CLASS_CONFIG to StringDeserializer::class.java,
-                    )
-
-                    // Create direct kafka stream with brokers and topics
-                    val messages: JavaInputDStream<ConsumerRecord<String, String>> = KafkaUtils.createDirectStream(
-                        ssc,
-                        LocationStrategies.PreferConsistent(),
-                        ConsumerStrategies.Subscribe(setOf(topic1, topic2), kafkaParams),
-                    )
-
-                    // Get the lines, split them into words, count the words and print
-                    val lines = messages.map { it.topic() X it.value() }
-                    val words = lines.flatMapValues { it.split(" ").iterator() }
-
-                    val wordCounts = words
-                        .map { t(it, 1) }
-                        .reduceByKey { a: Int, b: Int -> a + b }
-                        .map { (tup, counter) -> tup + counter }
-
-                    val resultLists = mapOf(
-                        topic1 to listOf(
-                            "Hello" X 1,
-                            "this" X 1,
-                            "is" X 1,
-                            "a" X 1,
-                            "test" X 3,
-                        ),
-                        topic2 to listOf(
-                            "This" X 1,
-                            "is" X 1,
-                            "also" X 2,
-                            "a" X 1,
-                            "test" X 2,
-                            "something" X 1,
-                        )
-                    )
-
-                    wordCounts.foreachRDD { rdd, _ ->
-                        rdd.foreach { (topic, word, count) ->
-                            t(word, count).shouldBeIn(collection = resultLists[topic]!!)
-                        }
-                    }
-
-                    wordCounts.print()
-                }
-            }
-
-        }
-
-
         should("stream") {
 
             val input = listOf("aaa", "bbb", "aaa", "ccc")
